completed tRex and Obstacle objects, we can get information about their status now.
